refactor: split reportes page into modular components
